[SDK21] Clean up Duration/Instant naming inconsistencies and add a few missing Java compat helpers

+++ urbanairship-automation/src/main/java/com/urbanairship/automation/engine/AutomationScheduleData.kt
@@ -246,7 +246,7 @@ internal class AutomationScheduleData(
             return false
         }
 
-        val gracePeriod = (schedule.editGracePeriodDays ?: return true).toLong().days
+        val gracePeriod = (schedule.editGracePeriodDays ?: return true).days
         val sinceLastChange = date - scheduleStateChangeDate
         return sinceLastChange >= gracePeriod
     }
